[GEOS-9487] rev 3
-moved test implemetnations inside unit test
-broke unit tests into three unit test
	-test geom generation
	-test re-projection
	-test cql filter
